Update SwiftPM version to 5.7
Also clean up what seems like a couple of mistakes from previous updates:
- remove a defunct `PDNextVersionLoadingTests.swift` (the new name is `PD_Next_LoadingTests.swift`)
- move a command plugin test to the 5.6 tests
