+++ Tests/PackageLoadingTests/PD_5_7_LoadingTests .swift
@@ -1,23 +1,20 @@
 /*
  This source file is part of the Swift.org open source project
 
- Copyright (c) 2021 Apple Inc. and the Swift project authors
+ Copyright (c) 2021 - 2022 Apple Inc. and the Swift project authors
  Licensed under Apache License v2.0 with Runtime Library Exception
 
  See http://swift.org/LICENSE.txt for license information
  See http://swift.org/CONTRIBUTORS.txt for Swift project authors
 */
 
+import Basics
+import PackageModel
+import SPMTestSupport
 import XCTest
 
-import TSCBasic
-import TSCUtility
-import SPMTestSupport
-import PackageModel
-import PackageLoading
-
-class PackageDescriptionNextVersionLoadingTests: PackageDescriptionLoadingTests {
+class PackageDescription5_7LoadingTests: PackageDescriptionLoadingTests {
     override var toolsVersion: ToolsVersion {
-        .vNext
+        .v5_7
     }
 }
